pyproject.toml
[project]
name = "xf-ai-agent"
version = "0.1.0"
description = "基于 LangChain、LangGraph 和 FastAPI 构建的智能代理后端项目"
authors = [
    { name = "lxf", email = "lxf@qq.com" }
]
dependencies = [
    "fastapi~=0.116.0",
    "uvicorn[standard]~=0.35.0",
    # LangChain 核心及生态包：最新版本 >= 1.2.10
    "langchain>=1.2.10",
    "langchain-core>=0.3.10",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-tavily>=0.0.1",
    "langgraph>=0.2.50", # LangGraph version sequence hasn't reached 1.2.x, their stable is around 0.2.x series, so requiring an update here.
    "langsmith>=0.1.130",
    # 环境变量加载
    "python-dotenv==1.1.0",
    # Tavily 搜索客户端（不变，与 langchain-tavily 配合使用）
    "tavily-python==0.7.3",
    # 文档处理相关
    "pypdf==3.10.0",
    # MCP 相关
    "langchain-mcp-adapters~=0.2.0", # 最新 0.2.1
    "mcp==1.12.1",
    # 网络请求
    "httpx==0.28.1",
    # Mysql相关
    "pymysql>=1.1.1",
    # 密码加密 - 使用兼容版本解决bcrypt版本读取警告
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    # token加密
    "pyjwt>=2.10.1",
    # 模板引擎
    "jinja2>=3.1.6",
    # MongoDB 相关
    "pymongo>=4.14.0",
    # 可视化
    "ipython>=9.4.0",
    # 图片处理
    "pillow>=11.3.0",
    # 缓存相关
    "redis>=6.4.0",
    # 可视化
    "streamlit>=1.49.1",
    # 密码加密
    "cryptography>=46.0.3",
    # 模型scope
    "modelscope>=1.31.0",
    "langgraph-cli[inmem]>=0.4.12",
    # 连接 PgSQL 的底层驱动
    "psycopg2-binary>=2.9.11",
    # 最新一代的 Postgres 驱动（也就是 psycopg3）
    "psycopg[binary]>=3.3.3",
    # LangGraph 官方的 PgSQL 记忆插件
    "langgraph-checkpoint-postgres>=3.0.4",
    "langchain-postgres>=0.0.17",
    "pgvector>=0.3.6",
    "dateparser>=1.3.0",
    "zhdate>=0.1",
    "alembic>=1.14.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
# uv 特定的配置，例如：
# resolver = "pip" # 可以指定解析器

[tool.setuptools.packages.find]
where = ["app"]
